refactor(ci): 启用 PR 全量验证与主线证据核对
- 将六项质量职责统一到受保护的 PR 可复用门禁并移除重复 push 触发
- 验证测试合并与主线提交的双亲和内容，以独立 App 绑定检查来源
- 采用批准的可信核心并保留同提交手动验证及发布凭据边界
