+++ .github/workflows/pr-quality-gate.yml
@@ -1,0 +1,20 @@
+name: Pull Request Quality Gate
+on:
+  pull_request:
+    branches:
+      - master
+    types:
+      - opened
+      - reopened
+      - synchronize
+      - edited
+permissions:
+  contents: read
+concurrency:
+  group: pr-quality-${{ github.event.pull_request.number }}-${{
+    github.event.action == 'edited' && github.event.changes.base == null }}
+  cancel-in-progress: true
+jobs:
+  quality-gate:
+    if: github.event.action != 'edited' || github.event.changes.base != null
+    uses: Sywyar/PixivDownloader/.github/workflows/quality-gate.yml@master
