--- .github/workflows/shared-snippets-check.yml
@@ -1,31 +1,0 @@
-name: Shared Snippet Drift Check
-
-# 守卫 scripts/shared/*.js 与各 .user.js 中 // >>> SHARED 标记区的一致性：
-# 若有人改了某一侧而未重新同步（mvn 构建会在 generate-sources 阶段调用
-# sync-shared-snippets.ps1 写回脚本），这里会在推送 / PR 时以非零退出码失败。
-on:
-  push:
-    branches: [master]
-  pull_request:
-  workflow_dispatch:
-
-permissions:
-  contents: read
-
-concurrency:
-  group: ${{ github.workflow }}-${{ github.event_name }}-${{ github.event.pull_request.number || github.run_id }}
-  cancel-in-progress: ${{ github.event_name == 'pull_request' }}
-
-jobs:
-  check-shared-snippets:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Fix Git UTF-8 filenames
-        run: git config --global core.quotepath false
-
-      - name: Checkout
-        uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1 # v7
-
-      - name: Verify .user.js shared snippets are in sync
-        shell: pwsh
-        run: ./scripts/sync-shared-snippets.ps1 -Check
